pyproject.toml
[tool.poetry]
name = "se-interview"
version = "0.1.0"
description = "Travel assistant assessment with LangGraph, FastAPI, and Arize Phoenix"
authors = ["Omar Ramadan <omar@example.com>"]
readme = "README.md"
packages = [{ include = "travel_assistant", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
langgraph = "^0.2.74"
langchain = "^0.3.23"
langchain-community = "^0.3.21"
langchain-openai = "^0.3.12"
ddgs = "^9.0"
python-dotenv = "^1.0"
fastapi = "^0.115"
uvicorn = "^0.32"
httpx = "^0.28.1"
pydantic = "^2.10.6"
arize-phoenix-client = "^1.31.0"
arize-phoenix-evals = "^2.11.0"
arize-phoenix-otel = "^0.14.0"
openinference-instrumentation-langchain = "^0.1.55"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["tests"]
